+++ app/models/user.rb
@@ -120,9 +120,6 @@ class User < Sequel::Model
 
   DEFAULT_MAX_LAYERS = 8
 
-  MIN_PASSWORD_LENGTH = 6
-  MAX_PASSWORD_LENGTH = 64
-
   GEOCODING_BLOCK_SIZE = 1000
   HERE_ISOLINES_BLOCK_SIZE = 1000
   OBS_SNAPSHOT_BLOCK_SIZE = 1000
@@ -235,25 +232,18 @@ class User < Sequel::Model
   end
 
   def valid_password?(key, value, confirmation_value)
-    if value.nil?
-      errors.add(key, "New password can't be blank")
-    else
-      if value != confirmation_value
-        errors.add(key, "New password doesn't match confirmation")
-      end
-
-      if value.length < MIN_PASSWORD_LENGTH
-        errors.add(key, "Must be at least #{MIN_PASSWORD_LENGTH} characters long")
-      end
-
-      if value.length >= MAX_PASSWORD_LENGTH
-        errors.add(key, "Must be at most #{MAX_PASSWORD_LENGTH} characters long")
-      end
-
-      validate_different_passwords(nil, self.class.password_digest(value, salt), key)
-    end
+    password_validator.validate(value, confirmation_value, self).each { |e| errors.add(key, e) }
+    validate_different_passwords(nil, self.class.password_digest(value, salt), key)
 
     errors[key].empty?
+  end
+
+  def password_validator
+    if organization.try(:strong_passwords_enabled)
+      Carto::PasswordValidator.new(Carto::StrongPasswordStrategy.new)
+    else
+      Carto::PasswordValidator.new(Carto::StandardPasswordStrategy.new)
+    end
   end
 
   def valid_creation?(creator_user)